Changed name of module Commented out java class requires metadata collection Fixed spacing issues
